Remove blank line from functions.php as requested by Austen:
From: Austen Frostad <Austen.Frostad@IntelliTect.com>
Date: Friday, July 10, 2020 at 12:09 PM
To: Erin Wissing <Erin.Wissing@IntelliTect.com>
Subject: Re: Site Map Issue - can you address this morning?

I fixed a file that most people say causes this issue (it is common) (https://wordpress.org/support/topic/sitemap-error-xml-declaration-allowed-only-at-the-start-of-the-document-2/ says to remove the blank line at the end of the functions.php file ). However the next step is to clear the website cache. I haven't done that, and want to make sure that doing that is okay. Could you forward this email to someone who might be able to do that?  Also the quick solution is to manually remove the blank line in the site_map.xml file, and this is done, by opening the file on the hosting service. Which I don't have access to.

//start of file
(blank line) <-remove this
<?xml version="1.0" encoding="UTF-8"?><?xml-stylesheet type="text/xsl" href="//intellitect.com/wp-content/plugins/wordpress-seo/css/main-sitemap.xsl"?>